TRIVIAL: FET-113 Move shared code to helper to avoid rerunning test cases
