+++ src/components/visualizations/chart/legend/test/legendBuilder.spec.ts
@@ -1,5 +1,6 @@
 // (C) 2007-2018 GoodData Corporation
-import { generateChartOptions } from '../../test/chartOptionsBuilder.spec';
+import { generateChartOptions } from '../../test/helper';
+
 import * as fixtures from '../../../../../../stories/test_data/fixtures';
 import getLegend, {
     shouldLegendBeEnabled,
